Fixed issue with request duplication

+++ server.py
@@ -19,6 +19,8 @@ class Server():
 
         # Initialize list for files that need to be uploaded across all servers
         self.upload_queue = []
+        self.rename_queue = []
+        self.delete_queue = []
 
         self.files_list = []
         self.get_files_list()
@@ -47,7 +49,7 @@ class Server():
             # Append client socket to list
             client_list.append(start_new_thread(self.on_new_client, (client_socket, address)))
 
-    def connect_to_slaves(self):
+    def connect_to_slaves(self, flag = True):
         # Wait for 5 seconds before connecting to peers
         time.sleep(5)
 
@@ -68,7 +70,8 @@ class Server():
 
                     # Keep synchronizing servers
                     # time.sleep(10)
-                    self.synchronize_servers()
+                    if flag:
+                        self.synchronize_servers()
 
                 except Exception as e:
                     print(e)
@@ -78,7 +81,7 @@ class Server():
                 self.synchronize_servers()
 
 
-    def download(self, client_socket, address, filename, flag=False):
+    def download(self, client_socket, address, filename):
         print(f"[*] Request to download {filename} from {address} received.")
         try:
             with open(self.SERVER_NAME + "/" + filename, "rb") as f:
@@ -94,7 +97,7 @@ class Server():
         except:
             print("[!] File not found.")
 
-    def upload(self, client_socket, address, filename, checksum=None):
+    def upload(self, client_socket, address, filename, checksum=None, flag=True):
         print(f"[*] Request to upload {filename} from {address} received.")
         try:
             if self.check(self.files_list, "checksum", checksum):
@@ -104,7 +107,7 @@ class Server():
                 client_socket.sendall(bytes(json.dumps(msg), encoding = "utf-8"))
                 pass
             else:
-                print("[*] File transfer started")
+                print("[*] File transfer started.")
                 # data = None
 
                 msg = {"command" : "ack"}
@@ -121,8 +124,14 @@ class Server():
                             break
                         f.write(bytes_read)
 
-                self.files_list.append({"filename" : filename, "checksum" : checksum})
-                self.upload_queue.append({"filename" : filename, "checksum" : checksum})
+                if flag:
+                    temp = {"filename" : filename, "checksum" : checksum}
+                    if temp not in self.files_list:
+                        self.files_list.append(temp)
+
+                    temp = {"filename" : filename, "checksum" : checksum}
+                    if temp not in self.upload_queue:
+                        self.upload_queue.append(temp)
 
             # print(self.files_list)
             # print(checksum)
@@ -134,6 +143,7 @@ class Server():
     def list_files(self, client_socket, address):
         print(f"[*] Request to get files list from {address} received.")
         try:
+            self.get_files_list()
             print("[*] Sending files list.")
 
             msg = {"files" : self.files_list}
@@ -144,35 +154,67 @@ class Server():
             print(e)
             print("[!] An error occured while sending files list.")
 
-    def delete_file(self, client_socket, address, filename):
+    def delete_file(self, client_socket, address, filename, checksum=None, flag=True):
         print(f"[*] Request to delete {filename} from {address} received.")
         try:
-            print("[*] Deleting file...")
-            os.remove(self.SERVER_NAME + "/" + filename)
+            if self.check(self.files_list, "checksum", checksum):
+                print("[!] File does not exist in the server.")
+                print("[-] Aborting file deletion.")
+                msg = {"command" : "nack"}
+                client_socket.sendall(bytes(json.dumps(msg), encoding = "utf-8"))
+                pass
+            else:
+                print("[*] File deletion started.")
+                # data = None
 
-            for file in self.files_list:
-                if(file["filename"] == filename):
-                    self.files_list.remove(file)
-                    break
+                msg = {"command" : "ack"}
+                client_socket.sendall(bytes(json.dumps(msg), encoding = "utf-8"))
 
-            print(f"[*] {filename} deleted.")
+                print("[*] Deleting file...")
+                os.remove(self.SERVER_NAME + "/" + filename)
+
+                for file in self.files_list:
+                    if(file["filename"] == filename):
+                        self.files_list.remove(file)
+                        temp = {"filename" : file["filename"], "checksum" : file["checksum"]}
+                        if temp not in self.delete_queue and flag:
+                            self.delete_queue.append(temp)
+                        break
+
+                print(f"[*] {filename} deleted.")
 
         except Exception as e:
             print(e)
             print("[!] An error occured while deleting file.")
 
-    def rename_file(self, client_socket, address, filename, newfilename):
+    def rename_file(self, client_socket, address, filename, newfilename, checksum=None, flag=True):
         print(f"[*] Request to rename {filename} to {newfilename} from {address} received.")
         try:
-            print("[*] Renaming file...")
-            os.rename(self.SERVER_NAME + "/" + filename, self.SERVER_NAME + "/" + newfilename)
+            if self.check(self.files_list, "checksum", checksum):
+                print("[!] File does not exist in the server.")
+                print("[-] Aborting file renaming.")
+                msg = {"command" : "nack"}
+                client_socket.sendall(bytes(json.dumps(msg), encoding = "utf-8"))
+                pass
+            else:
+                print("[*] File renaming started.")
+                # data = None
 
-            for file in self.files_list:
-                if(file["filename"] == filename):
-                    file["filename"] = newfilename
-                    break
+                msg = {"command" : "ack"}
+                client_socket.sendall(bytes(json.dumps(msg), encoding = "utf-8"))
 
-            print(f"[*] {filename} renamed to {newfilename}.")
+                print("[*] Renaming file...")
+                os.rename(self.SERVER_NAME + "/" + filename, self.SERVER_NAME + "/" + newfilename)
+
+                for file in self.files_list:
+                    if(file["filename"] == filename):
+                        file["filename"] = newfilename
+                        temp = {"filename" : filename, "newfilename" : newfilename, "checksum" : file["checksum"]}
+                        if temp not in self.rename_queue and flag:
+                            self.rename_queue.append(temp)
+                        break
+
+                print(f"[*] {filename} renamed to {newfilename}.")
 
         except Exception as e:
             print(e)
@@ -183,10 +225,10 @@ class Server():
         try:
             if self.upload_queue:
                 for file in self.upload_queue:
-                    print("[*] New file was uploaded. Synchronizing servers.")
+                    print("[*] A new file was uploaded. Synchronizing servers.")
                     for peer in self.server_sockets:
 
-                        msg = {"command" : "upload", "filename" : file["filename"], "checksum" : file["checksum"]}
+                        msg = {"command" : "upload", "filename" : file["filename"], "checksum" : file["checksum"], "flag" : False}
 
                         peer.sendall(bytes(json.dumps(msg), encoding = "utf-8"))
 
@@ -196,7 +238,7 @@ class Server():
                                 parsed = json.loads(message)
                                 # print(str(parsed))
                                 if parsed["command"] == "ack":
-                                    self.download(peer, "127.0.0.1", file["filename"])
+                                    self.download(peer, "127.0.0.1", file["filename"], False)
                                     break
 
                                 elif parsed["command"] == "nack":
@@ -206,11 +248,71 @@ class Server():
                     self.upload_queue.remove(file)
                     print("[*] Re-establishing connection with slaves...")
                     self.server_sockets = []
+
+            if self.server_sockets == []:
+                self.connect_to_slaves(False)
+
+            if self.rename_queue:
+                for file in self.rename_queue:
+                    print("[*] A file was renamed. Synchronizing servers.")
+                    for peer in self.server_sockets:
+
+                        msg = {"command" : "rename", "filename" : file["filename"], "newfilename" : file["newfilename"], "checksum" : file["checksum"], "flag" : False}
+
+                        peer.sendall(bytes(json.dumps(msg), encoding = "utf-8"))
+
+                        while True:
+                            message = peer.recv(self.BUFFER_SIZE).decode("utf-8")
+                            if message:
+                                parsed = json.loads(message)
+                                # print(str(parsed))
+                                if parsed["command"] == "ack":
+                                    # self.rename_file(peer, "127.0.0.1", file["filename"], file["newfilename"], file["checksum"], False)
+                                    break
+
+                                elif parsed["command"] == "nack":
+                                    print("[!] File does not exist in the server.")
+                                    break
+                        peer.close()
+                    self.rename_queue.remove(file)
+                    print("[*] Re-establishing connection with slaves...")
+                    self.server_sockets = []
+
+            if self.server_sockets == []:
+                self.connect_to_slaves(False)
+
+            if self.delete_queue:
+                for file in self.delete_queue:
+                    print("[*] A file was deleted. Synchronizing servers.")
+                    for peer in self.server_sockets:
+
+                        msg = {"command" : "delete", "filename" : file["filename"], "checksum" : file["checksum"], "flag" : False}
+
+                        peer.sendall(bytes(json.dumps(msg), encoding = "utf-8"))
+
+                        while True:
+                            message = peer.recv(self.BUFFER_SIZE).decode("utf-8")
+                            if message:
+                                parsed = json.loads(message)
+                                # print(str(parsed))
+                                if parsed["command"] == "ack":
+                                    # self.delete_file(peer, "127.0.0.1", file["filename"], file["checksum"], False)
+                                    break
+
+                                elif parsed["command"] == "nack":
+                                    print("[!] File does not exist in the server.")
+                                    break
+                        peer.close()
+                    self.delete_queue.remove(file)
+                    print("[*] Re-establishing connection with slaves...")
+                    self.server_sockets = []
+
         except Exception as e:
             print(e)
             pass
 
     def get_files_list(self):
+        self.files_list = []
         for root, dir, files in os.walk(self.SERVER_NAME + "/"):
             for file in files:
                 data = open(self.SERVER_NAME + "/" + file, "rb").read()
@@ -227,25 +329,27 @@ class Server():
                     # print(str(parsed))
                     if parsed["command"] == "download":
                         filename = parsed["filename"]
-                        flag = parsed["flag"]
-                        self.download(client_socket, address, filename, flag)
+                        self.download(client_socket, address, filename)
 
                     elif parsed["command"] == "upload":
                         filename = parsed["filename"]
                         checksum = parsed["checksum"]
-                        self.upload(client_socket, address, filename, checksum)
+                        flag = parsed["flag"]
+                        self.upload(client_socket, address, filename, checksum, flag)
 
                     elif parsed["command"] == "list":
                         self.list_files(client_socket, address)
 
                     elif parsed["command"] == "delete":
                         filename = parsed["filename"]
-                        self.delete_file(client_socket, address, filename)
+                        flag = parsed["flag"]
+                        self.delete_file(client_socket, address, filename, None, flag)
 
                     elif parsed["command"] == "rename":
                         filename = parsed["filename"]
                         newfilename = parsed["newfilename"]
-                        self.rename_file(client_socket, address, filename, newfilename)
+                        flag = parsed["flag"]
+                        self.rename_file(client_socket, address, filename, newfilename, None, flag)
 
                     elif parsed["command"] == "exit":
                         client_socket.close()
